+++ src/RecoveryServices/RecoveryServices/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added support for Multi-user authorization using Resource Guard for recovery sevrices vault.
+* Added support for cross subscription restore for recovery services vault.
 
 ## Version 5.3.1
 * Added support for multiple backups per day (hourly) Enhanced policy for workloadType AzureVM.